Move utils pos, rename mem_tracker, update utils

+++ scripts/train_flux_pref.py
@@ -30,7 +30,7 @@ from peft import LoraConfig, get_peft_model, set_peft_model_state_dict, PeftMode
 from PIL import Image
 from torch.utils.data import Dataset, DataLoader, Sampler
 
-from flow_grpo.rewards.utils import tensor_to_pil_image, numpy_to_pil_image, numpy_list_to_pil_image, tensor_list_to_pil_image
+from flow_grpo.utils import tensor_to_pil_image, numpy_to_pil_image, numpy_list_to_pil_image, tensor_list_to_pil_image, gather_tensor_list
 from flow_grpo.rewards.rewards import multi_score
 from flow_grpo.rewards.pref_scorer import pref_score
 from flow_grpo.diffusers_patch.flux_pipeline_flexible_with_logprob import calculate_shift, pipeline_with_logprob, denoising_sde_step_with_logprob, compute_log_prob
@@ -40,7 +40,7 @@ from flow_grpo.stat_tracking import PerPromptStatTracker
 from flow_grpo.datasets.prompt_dataset import TextPromptDataset, GenevalPromptDataset
 from flow_grpo.datasets.sampler import DistributedKRepeatSampler
 from flow_grpo.scheduler import FlowMatchSlidingWindowScheduler
-from flow_grpo.debug_utils import MemoryProfiler
+from flow_grpo.memory_tracker import MemoryProfiler
 
 tqdm = partial(tqdm.tqdm, dynamic_ncols=True)
 
@@ -85,6 +85,7 @@ def eval(pipeline : FluxPipeline,
          autocast,
          ema,
          transformer_trainable_parameters,
+         memory_profiler : Optional[MemoryProfiler] = None,
          log_sample_num : int = 90 # 108 as max in wandb/swanlab
     ):
     if config.train.ema:
@@ -95,12 +96,17 @@ def eval(pipeline : FluxPipeline,
         'prompts': [],
         'rewards': defaultdict(list)
     }
-    for test_batch in tqdm(
-            test_dataloader,
+    if memory_profiler is not None:
+        memory_profiler.snapshot("before_eval")
+
+    for i, test_batch in tqdm(
+            enumerate(test_dataloader),
             desc="Eval: ",
             disable=not accelerator.is_local_main_process,
             position=0,
         ):
+        if memory_profiler is not None:
+            memory_profiler.snapshot(f"eval_batch_{i}_start")
 
         prompts, prompt_metadata = test_batch
         prompt_embeds, pooled_prompt_embeds = compute_text_embeddings(
@@ -154,24 +160,30 @@ def eval(pipeline : FluxPipeline,
                     width=widths[0],
                     noise_level=0,
                 )
+                images = list(images.unbind(0)) # List[torch.Tensor(C, H, W)]
         # reward_fn accepts torch.Tensor (B, C, H, W) or List[torch.Tensor(C, H, W)]
         future = executor.submit(reward_fn, images, prompts, prompt_metadata)
         # yield to to make sure reward computation starts
         time.sleep(0)
         # all_futures.append(future)
         rewards, reward_metadata = future.result()
-        
+    
         # -------------------------------Collect log data--------------------------------
         if len(log_data["prompts"]) < log_sample_num // accelerator.num_processes:
-            log_data['images'].extend([img.cpu().numpy() for img in images])
+            log_data['images'].extend(images)
             log_data['prompts'].extend(prompts)
             for key, value in rewards.items():
                 if key not in log_data['rewards']:
                     log_data['rewards'][key] = []
                 
                 log_data['rewards'][key].extend(value)
+        
+        # log memory after reward computation
+        if memory_profiler is not None:
+            memory_profiler.snapshot(f"eval_batch_{i}_end")
 
-
+    if memory_profiler is not None:
+        memory_profiler.snapshot("after_eval_before_gather_log_data")
     # ---------------------------Gather all Log data, with prompt-image-reward tuples--------------------------
     # 1. Gather all rewards and report average
     gathered_rewards = {}
@@ -197,6 +209,9 @@ def eval(pipeline : FluxPipeline,
         for value in zip(*gathered_rewards.values())
     ]
 
+    if memory_profiler is not None:
+        memory_profiler.snapshot("after_gather_rewards")
+
     # 2. Encode prompt to tensors for gpu communication
     prompt_ids = tokenizers[1](
         log_data['prompts'],
@@ -210,20 +225,33 @@ def eval(pipeline : FluxPipeline,
         gathered_prompt_ids, skip_special_tokens=True
     )
 
+    if memory_profiler is not None:
+        memory_profiler.snapshot("after_gather_prompts")
+
     # 3. Gather all images
+    use_jpg_compression = True
     # Approach : by saving them in a temp dir
-    temp_dir = os.path.join(config.save_dir, 'temp_eval_images')
-    os.makedirs(temp_dir, exist_ok=True)
-    for i,img in enumerate(log_data['images']):
-        # Save image to temp dir
-        img_id = f"{accelerator.process_index * len(log_data['images']) + i}.jpg"
-        pil = Image.fromarray((img.transpose(1, 2, 0) * 255).astype(np.uint8))
-        pil.save(os.path.join(temp_dir, img_id))
-    
-    accelerator.wait_for_everyone()
-    # The order of images here should be guaranteed by the name of images
-    # NOTE: it provides gathered_images as a list of file paths
-    gathered_images = [os.path.join(temp_dir, f) for f in sorted(os.listdir(temp_dir), key=lambda x: int(x.split('.')[0]))]
+    if use_jpg_compression:
+        # This approach saves images as JPG files in a temporary directory
+        # Since uploading images with jpg is faster, if we need to do it anyway.
+        temp_dir = os.path.join(config.save_dir, 'temp_eval_images')
+        os.makedirs(temp_dir, exist_ok=True)
+        offset = accelerator.process_index * len(log_data['images'])
+        for i,img in enumerate(log_data['images']):
+            # Save image to temp dir
+            pil_img = tensor_to_pil_image(img)[0]
+            pil_img.save(os.path.join(temp_dir, f"{offset + i}.jpg"))
+        accelerator.wait_for_everyone()
+        # The order of images here should be guaranteed by the name of images
+        # NOTE: it provides gathered_images as a list of file paths
+        gathered_images = [os.path.join(temp_dir, f) for f in sorted(os.listdir(temp_dir), key=lambda x: int(x.split('.')[0]))]
+    else:
+        # Approach: flatten and gather, then reshape
+        gathered_images = gather_tensor_list(accelerator, log_data['images'], device="cpu")
+        gathered_images = tensor_list_to_pil_image(gathered_images) # List[PIL.Image]
+
+    if memory_profiler is not None:
+        memory_profiler.snapshot("after_gather_images")
 
     if accelerator.is_main_process:
          # Use a fixed generator to log same indices everytime for comparison
@@ -251,10 +279,15 @@ def eval(pipeline : FluxPipeline,
             step=global_step,
         )
         # Clean up temp dir
-        shutil.rmtree(temp_dir)
+        if use_jpg_compression:
+            shutil.rmtree(temp_dir)
 
     if config.train.ema:
         ema.copy_temp_to(transformer_trainable_parameters)
+
+    # Log memory after eval
+    if memory_profiler is not None:
+        memory_profiler.snapshot("after_eval")
 
 def unwrap_model(model, accelerator):
     model = accelerator.unwrap_model(model)
@@ -718,7 +751,8 @@ These two numbers should be equal
                 eval_reward_fn, executor,
                 autocast,
                 ema,
-                transformer_trainable_parameters
+                transformer_trainable_parameters,
+                memory_profiler=memory_profiler if config.enable_mem_log else None,
             )
             if config.enable_mem_log:
                 memory_profiler.snapshot(f"epoch_{epoch}_after_eval")
